rm duplicate script + fixes
Fix:
* no need to exec a shell with chroot command.
* mv was unnecessary since the files get saved outside the chroot by
  default.

+++ scripts/genfiles.sh
@@ -7,23 +7,20 @@ DIR="$1"
 SET="$2"
 
 # Sets get saved in the / directory. Move it in current dir first.
-mv /"$SET" .
+[ -f /"$SET" ] && mv /"$SET" .
 
 # Create target dir and extract the set
 mkdir -p "$DIR"
 tar -xpvf "$SET" --xattrs --xattrs-include='*' -C "$DIR"
 
 # Create the .contents file
-chroot "$DIR" /bin/ksh find . | sed -e 's|^.||g' | grep "^/" | sort > "$SET".contents
-mv "$DIR"/"$SET".contents .
+chroot "$DIR" find . | sed -e 's|^.||g' | grep "^/" | sort > "$SET".contents
 
 # Create the .files file
-chroot "$DIR" /bin/ksh ls -laR > "$SET".files
-mv "$DIR"/"$SET".files .
+chroot "$DIR" ls -laR > "$SET".files
 
 # Create the .packages file
-chroot "$DIR" /bin/ksh dpkg-query -W > "$SET".packages
-mv "$DIR"/"$SET".packages .
+chroot "$DIR" dpkg-query -W > "$SET".packages
 
 # Delete the extracted set
 rm -r "$DIR"
